refactor(studio): rename 'library' components to 'releases'
