+++ apps/studio/src/components/index.ts
@@ -10,8 +10,8 @@ export { default as DisconnectWalletButton } from "./DisconnectWalletButton";
 export { default as DistributionPricingDialog } from "./dspPricing/DistributionPricingDialog";
 export { default as AppleLogin } from "./AppleLogin";
 export { default as GoogleLogin } from "./GoogleLogin";
-export * from "./library/IconStatus";
-export { default as IconStatus } from "./library/IconStatus";
+export * from "./releases/IconStatus";
+export { default as IconStatus } from "./releases/IconStatus";
 export * from "./idenfy";
 export { default as InvitesModal } from "./invites/InvitesModal";
 export { default as LogoutButton } from "./home/LogoutButton";
@@ -35,7 +35,7 @@ export { default as ReferralDashboard } from "./referralHero/ReferralDashboard";
 export { default as ResponsiveNEWMLogo } from "./ResponsiveNEWMLogo";
 export { SearchBox } from "./SearchBox";
 export { default as SelectWalletItem } from "./uploadSong/SelectWalletItem";
-export { default as SongStreamPlaybackIcon } from "./library/SongStreamPlaybackIcon";
+export { default as SongStreamPlaybackIcon } from "./releases/SongStreamPlaybackIcon";
 export { default as SquareGridCard } from "./SquareGridCard";
 export * from "./TableDropdownSelect";
 export { default as TableDropdownSelect } from "./TableDropdownSelect";
